backend/pyproject.toml
[project]
name = "summaries-website"
version = "0.1.0"
description = "Backend Website for summaries"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "commit-check>=0.9.7",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-pythonpath>=0.7.3",
    "ruff>=0.12.0",
    "ty>=0.0.1a11",
]
